Proper User-Character Relationship 3/3 - Delete & Get Single Character

+++ Services/CharacterService/CharacterService.cs
@@ -63,11 +63,20 @@ namespace first_web_api.Services.CharacterService
                 //Character character = Test.First(c => c.Id == id);
                 //Test.Remove(character);
                 //serviceResponse.Data = Test.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList(); // if i put await and toListASync will not remove it because it provides search sort ..etc not delete
-
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id==GetUserID());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserID())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList(); // if i put await and toListASync will not remove it because it provides search sort ..etc not delete
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Messsage = "Charachter Not Found "!;
+                }
             }
             catch (Exception ex)
             {
@@ -85,7 +94,6 @@ namespace first_web_api.Services.CharacterService
             var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserID()).ToListAsync();
             //serviceResponse.Data = Test.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
-
             return serviceResponse;
         }
 
@@ -94,7 +102,7 @@ namespace first_web_api.Services.CharacterService
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id==GetUserID());
                 //serviceResponse.Data = _mapper.Map< GetCharacterDto >(Test.FirstOrDefault(c => c.Id == id));
                 serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             }
